// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        VendorOut: {
            /** Vendor Name */
            vendor_name: string;
            /** Pan */
            pan?: string | null;
            /** Gstin */
            gstin?: string | null;
            /**
             * Default Tds Slab Pct
             * @default 2.00
             */
            default_tds_slab_pct: string;
            /**
             * Category
             * @default contractor
             * @enum {string}
             */
            category: "contractor" | "professional";
            /**
             * Default Payment Terms Days
             * @default 45
             */
            default_payment_terms_days: number;
            /** Id */
            id: number;
            /** Is Active */
            is_active: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}